Fix 403

+++ Ease/src/com/Ease/API/V1/Teams/ServletDeleteSharedApp.java
@@ -4,6 +4,9 @@ import com.Ease.Dashboard.App.ShareableApp;
 import com.Ease.Dashboard.App.SharedApp;
 import com.Ease.Team.Team;
 import com.Ease.Team.TeamManager;
+import com.Ease.Team.TeamUser;
+import com.Ease.Utils.HttpServletException;
+import com.Ease.Utils.HttpStatus;
 import com.Ease.Utils.Servlets.PostServletManager;
 import com.Ease.websocketV1.WebSocketMessageAction;
 import com.Ease.websocketV1.WebSocketMessageFactory;
@@ -27,11 +30,14 @@ public class ServletDeleteSharedApp extends HttpServlet {
         PostServletManager sm = new PostServletManager(this.getClass().getName(), request, response, true);
         try {
             Integer team_id = sm.getIntParam("team_id", true, false);
-            sm.needToBeAdminOfTeam(team_id);
+            sm.needToBeTeamUserOfTeam(team_id);
             TeamManager teamManager = (TeamManager) sm.getContextAttr("teamManager");
             Team team = teamManager.getTeamWithId(team_id);
+            TeamUser teamUser_connected = sm.getTeamUserForTeam(team);
             Integer sharedApp_id = sm.getIntParam("shared_app_id", true, false);
             SharedApp sharedApp = team.getAppManager().getSharedApp(sharedApp_id);
+            if (sharedApp.getTeamUser_tenant() != teamUser_connected && !teamUser_connected.isTeamAdmin())
+                throw new HttpServletException(HttpStatus.Forbidden);
             ShareableApp shareableApp = sharedApp.getHolder();
             team.getAppManager().removeSharedApp(sharedApp);
             sharedApp.deleteShared(sm.getDB());
